#2005: extract forward/reverse tuple match into session/lookup.rs
Pure code-motion split of the conntrack read path. Move the lookup
family off SessionTable's monolithic impl in session/mod.rs into a
focused session/lookup.rs submodule: lookup, lookup_with_origin,
find_forward_nat_match, find_forward_wire_match,
find_forward_wire_match_with_origin, entry_with_origin,
owner_rg_session_keys, take_synced_local, iter_with_origin,
iter_with_idle, iter_with_idle_and_origin. Bodies are byte-for-byte
identical; only the file boundary changed.

No visibility was widened: every moved method is already `pub` (callable
crate-wide off the pub(crate) SessionTable type), and the private
helpers they call — entry_by_key, remove_entry, push_to_wheel (now
pub(in crate::session)), session_timeout_ns, and the wire-key
transforms in key.rs — live in ancestor/sibling modules and stay
visible to this descendant. lookup_with_origin uses the debug_log!
macro, so `mod lookup;` sits after the macro definition (textual
scope).

The in-place-refresh contract (#1752/#1855) and the secondary-index /
delta-queue ownership stay in mod.rs; demote_owner_rg,
drain_deltas/has_pending_deltas, and push_delta were left in place.

Validation: cargo build --release clean; cargo test --release session::
78 session tests pass.
